feat: moved NatCheckboxContainer to its own file
